requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.1
python-dotenv==1.0.1
requests==2.32.3
scratchattach==1.7.3
soupsieve==2.5
urllib3==2.2.2
websocket-client==1.8.0
